fix: auth 페이지 분리
